Create new file to camera manager
